fix(review): move campaign tests to test_fleet_campaign.py; strengthen assertion
Move test_fleet_campaign_greetings_have_campaign_name_placeholder,
test_launch_fleet_session_forwards_initial_message_campaign, and
test_launch_fleet_session_clears_initial_message_on_reload_campaign from
test_fleet_dispatch.py to test_fleet_campaign.py where they belong.

Strengthen test_launch_fleet_session_forwards_initial_message_campaign to
capture per-call kwargs list and assert exactly one call was made, preventing
silent pass if initial_message were cleared on a second iteration.
